make new performance test class
